kvmconfig: extract metric package
This prevents future panics when copying packages.
