+++ service/kvmconfig/v4/resource/deployment/master_deployment.go
@@ -72,7 +72,7 @@ func newMasterDeployments(customObject v1alpha1.KVMConfig) ([]*extensionsv1.Depl
 			ObjectMeta: apismetav1.ObjectMeta{
 				Name: key.DeploymentName(key.MasterID, masterNode.ID),
 				Annotations: map[string]string{
-					VersionBundleVersionAnnotation: key.VersionBundleVersion(customObject),
+					key.VersionBundleVersionAnnotation: key.VersionBundleVersion(customObject),
 				},
 				Labels: map[string]string{
 					"cluster":  key.ClusterID(customObject),
